Stokhos:  Modify MP::Vector to allow constexpr constructors.
This allows MP::Vector to be used in constexpr expressions of the form
   static constexpr MP::Vector<..> x = ...;
This would have been easier with C++14, and even easier with C++17.

Don't merge this without further investigation.  It causes linking
errors with Belos's default values, I think because how they are passed
into this constructor makes them odr-used.  This I think indicates that
this approach is not very efficient, since it is recursive on the vector
length, which probably runs into an inline/optimization limit.  Maybe
wait for C++17, or find a better way to do this.